requirements.txt
oryks-youtube==0.10.0
chainlit==0.7.700
thefuzz==0.20.0